Removed ip_allow.config, remap.config, and parent.config v1 generation routes
